requirements.txt
# build==0.1.0
matplotlib==3.1.2
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
# osrf-pycommon==2.0.2
pandas==1.3.4
Pillow==7.0.0
# pymap3d==3.0.1
pyproj==3.5.0
PyQt5-sip==12.12.1
PyQt5==5.15.4
python-qt-binding==0.4.4
PyYAML==5.3.1
rasterio==1.2.1
scikit-learn==0.24.2
scipy==1.10.1
setuptools
tk==0.1.0